consolidate buttons in Toolbar.tsx
