--- catalog/app/containers/Bucket/Toolbar/Organize.tsx
@@ -1,30 +1,0 @@
-import * as React from 'react'
-import * as M from '@material-ui/core'
-
-import { PlaylistAddCheckOutlined as IconPlaylistAddCheckOutlined } from '@material-ui/icons'
-
-import * as Buttons from 'components/Buttons'
-import { useSelection } from 'containers/Bucket/Selection/Provider'
-
-const useBadgeClasses = M.makeStyles({
-  badge: {
-    right: '4px',
-  },
-})
-
-type ButtonProps = Omit<Buttons.WithPopoverProps, 'icon'>
-
-export default function Button(props: ButtonProps) {
-  const slt = useSelection()
-  const classes = useBadgeClasses()
-  return (
-    <M.Badge badgeContent={slt.totalCount} classes={classes} color="primary" max={999}>
-      <Buttons.WithPopover
-        icon={IconPlaylistAddCheckOutlined}
-        label="Organize"
-        disabled={slt.inited && !slt.totalCount}
-        {...props}
-      />
-    </M.Badge>
-  )
-}
